requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-admin-sortable==2.3
django-cors-headers==3.14.0
django-graphql-jwt==0.3.4
graphene==3.2.2
graphene-django==3.0.0
graphene-file-upload==1.3.0
graphql-core==3.2.3
graphql-relay==3.2.0
promise==2.3
pybase62==0.6.0
PyJWT==2.6.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
